replacing uint literal with golangs syntax

+++ pkg/libwebp/dec/buffer_dec.c.go
@@ -109,7 +109,7 @@ static VP8StatusCode AllocateBuffer(*WebPDecBuffer const buffer) {
     int stride;
     uint64 size;
 
-    if ((uint64)w * kModeBpp[mode] >= (1ull << 31)) {
+    if ((uint64)w * kModeBpp[mode] >= (uint64(1) << 31)) {
       return VP8_STATUS_INVALID_PARAM;
     }
     stride = w * kModeBpp[mode];
